Changed name to better describe function. Add and delete tests are passing.
